backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
httpx>=0.26.0
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pyotp
qrcode
minio

